Refaktoryzacja, dodanie testow stanu
