+++ src/test/java/pl/com/bottega/ecommerce/sales/domain/invoicing/ProductBuilder.java
@@ -1,0 +1,38 @@
+package pl.com.bottega.ecommerce.sales.domain.invoicing;
+
+import pl.com.bottega.ecommerce.canonicalmodel.publishedlanguage.Id;
+import pl.com.bottega.ecommerce.sales.domain.productscatalog.Product;
+import pl.com.bottega.ecommerce.sales.domain.productscatalog.ProductType;
+import pl.com.bottega.ecommerce.sharedkernel.Money;
+
+public class ProductBuilder {
+    private Id aggregateId = Id.generate();
+    private Money price = Money.ZERO;
+    private String name = "ProductName";
+    private ProductType productType = ProductType.STANDARD;
+
+    public ProductBuilder withAggregateId(Id aggregateId) {
+        this.aggregateId = aggregateId;
+        return this;
+    }
+
+    public ProductBuilder withPrice(Money price) {
+        this.price = price;
+        return this;
+    }
+
+    public ProductBuilder withName(String name) {
+        this.name = name;
+        return this;
+    }
+
+    public ProductBuilder withProductType(ProductType productType) {
+        this.productType = productType;
+        return this;
+    }
+
+    public Product build() {
+        return new Product(aggregateId, price, name, productType);
+    }
+}
+
